Split up onboarding code (#3158)
