+++ src/data/onboarding.ts
@@ -1,12 +1,19 @@
 import { handleFetchPromise } from "../util/hass-call-api";
 
-export interface OnboardingStep {
-  step: string;
-  done: boolean;
+export interface OnboardingUserStepResponse {
+  auth_code: string;
 }
 
-interface UserStepResponse {
-  auth_code: string;
+export interface OnboardingResponses {
+  user: OnboardingUserStepResponse;
+  bla: number;
+}
+
+export type ValidOnboardingStep = keyof OnboardingResponses;
+
+export interface OnboardingStep {
+  step: ValidOnboardingStep;
+  done: boolean;
 }
 
 export const fetchOnboardingOverview = () =>
@@ -18,7 +25,7 @@ export const onboardUserStep = (params: {
   username: string;
   password: string;
 }) =>
-  handleFetchPromise<UserStepResponse>(
+  handleFetchPromise<OnboardingUserStepResponse>(
     fetch("/api/onboarding/users", {
       method: "POST",
       credentials: "same-origin",
